Pull Error banner out to its own file
